requirements.txt
python-telegram-bot
langchain
langchain-community
openai
tiktoken
google-generativeai 
langchain-google-genai
sentence-transformers
langchain-google-vertexai
llama_index
llama-index-embeddings-langchain
faiss-cpu